DashboardMicroservice/requirements.txt
Flask==3.0.2
nats-py
numpy
influxdb_client
python-dotenv